requirements.txt
pandas~=2.0.1
kubernetes~=26.1.0
pendulum~=2.1.2
python-dateutil~=2.8.2
sqlparse~=0.4.4
pytz~=2023.3
PyMySQL==1.1.0
scikit-learn==1.2.2
joblib~=1.3.1
SQLAlchemy==1.4.47
